Agent: Move UI Events Into App Domain
Move application-wide event contracts from app/components to app/events because they coordinate UI surfaces rather than implement a component.
