+++ src/ui/src/app/components/gm-live-reload-panel.ts
@@ -8,13 +8,13 @@ import type {
     GraphVisualizationLiveReloadWatcherStatus
 } from "../../graph/index.js";
 import type { GraphVisualizationUiModel } from "../contracts.js";
-import type { GraphVisualizationUiState } from "../state/types.js";
-import { EventBusManager } from "./event-bus-mixin.js";
 import {
     GRAPH_UI_EVENT_CLEAR_PAGE_ERROR,
     GRAPH_UI_EVENT_LIVE_RELOAD_STATUS_CHANGED,
     type GraphUiLiveReloadStatusChangedDetail
-} from "./events.js";
+} from "../events/events.js";
+import type { GraphVisualizationUiState } from "../state/types.js";
+import { EventBusManager } from "./event-bus-mixin.js";
 import { LifecycleParticipantsController } from "./lifecycle-participants-controller.js";
 import { LightDomLitElement } from "./light-dom-lit-element.js";
 import { LiveReloadPollingController } from "./live-reload-polling-controller.js";
