+++ src/ui/src/app/components/index.ts
@@ -1,7 +1,6 @@
 export * from "./copy-clipboard.js";
 export * from "./copy-feedback-controller.js";
 export * from "./event-bus-mixin.js";
-export * from "./events.js";
 export * from "./fix-workflow-reconnect-participant.js";
 export * from "./gm-app-header.js";
 export * from "./gm-app-shell.js";
