+++ src/ui/src/app/index.ts
@@ -1,4 +1,5 @@
 export * from "./bootstrap.js";
 export * as Components from "./components/index.js";
 export * from "./contracts.js";
+export * as Events from "./events/index.js";
 export * as State from "./state/index.js";
