+++ src/ui/test/fix-cancel-workflow.test.ts
@@ -3,9 +3,9 @@ import test from "node:test";
 
 import type { PropertyValues } from "lit";
 
-import { GRAPH_UI_EVENT_TRIGGER_CANCEL_FIX } from "../src/app/components/events.js";
 import { GmAppShell } from "../src/app/components/gm-app-shell.js";
 import { createNoopGraphVisualizationUiCallbacks, type GraphVisualizationUiModel } from "../src/app/contracts.js";
+import { GRAPH_UI_EVENT_TRIGGER_CANCEL_FIX } from "../src/app/events/events.js";
 
 class TestableGmAppShell extends GmAppShell {
     // Tests drive the store directly and never render the shell into a real
